# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Owlin API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/upload/enhanced": {
      "post": {
        "tags": [
          "enhanced-upload"
        ],
        "summary": "Upload Document Enhanced",
        "description": "Enhanced upload endpoint with adaptive processing and comprehensive error recovery\n\nArgs:\n    file: Uploaded file\n    parse_templates: Whether to parse invoice templates\n    save_debug: Whether to save debug artifacts\n    user_role: User role for role-based processing\n    \nReturns:\n    UploadResponse with processing results",
        "operationId": "upload_document_enhanced_api_upload_enhanced_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_enhanced_api_upload_enhanced_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/upload/adaptive": {
      "post": {
        "tags": [
          "enhanced-upload"
        ],
        "summary": "Upload Document Adaptive",
        "description": "Adaptive upload endpoint with progress tracking and timeout handling\n\nArgs:\n    file: Uploaded file\n    parse_templates: Whether to parse invoice templates\n    save_debug: Whether to save debug artifacts\n    user_role: User role for role-based processing\n    \nReturns:\n    UploadResponse with processing results",
        "operationId": "upload_document_adaptive_api_upload_adaptive_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_adaptive_api_upload_adaptive_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/upload/batch": {
      "post": {
        "tags": [
          "enhanced-upload"
        ],
        "summary": "Upload Batch Enhanced",
        "description": "Enhanced batch upload endpoint for multiple documents\n\nArgs:\n    files: List of uploaded files\n    parse_templates: Whether to parse invoice templates\n    save_debug: Whether to save debug artifacts\n    user_role: User role for role-based processing\n    \nReturns:\n    Dictionary with batch processing results",
        "operationId": "upload_batch_enhanced_api_upload_batch_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_batch_enhanced_api_upload_batch_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Upload Batch Enhanced Api Upload Batch Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/upload/status/{document_id}": {
      "get": {
        "tags": [
          "enhanced-upload"
        ],
        "summary": "Get Upload Status",
        "description": "Get processing status for a document\n\nArgs:\n    document_id: Document ID\n    \nReturns:\n    ProcessingStatus with current status",
        "operationId": "get_upload_status_api_upload_status__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/upload/{document_id}": {
      "delete": {
        "tags": [
          "enhanced-upload"
        ],
        "summary": "Delete Document",
        "description": "Delete a processed document\n\nArgs:\n    document_id: Document ID\n    \nReturns:\n    Success status",
        "operationId": "delete_document_api_upload__document_id__delete",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Document Api Upload  Document Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_batch_enhanced_api_upload_batch_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "format": "binary"
            },
            "type": "array",
            "title": "Files"
          },
          "parse_templates": {
            "type": "boolean",
            "title": "Parse Templates",
            "default": true
          },
          "save_debug": {
            "type": "boolean",
            "title": "Save Debug",
            "default": false
          },
          "user_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Role"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_batch_enhanced_api_upload_batch_post"
      },
...
      "Body_upload_document_adaptive_api_upload_adaptive_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "parse_templates": {
            "type": "boolean",
            "title": "Parse Templates",
            "default": true
          },
          "save_debug": {
            "type": "boolean",
            "title": "Save Debug",
            "default": false
          },
          "user_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Role"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_adaptive_api_upload_adaptive_post"
      },
      "Body_upload_document_enhanced_api_upload_enhanced_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "parse_templates": {
            "type": "boolean",
            "title": "Parse Templates",
            "default": true
          },
          "save_debug": {
            "type": "boolean",
            "title": "Save Debug",
            "default": false
          },
          "user_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Role"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_enhanced_api_upload_enhanced_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProcessingStatus": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress_percentage": {
            "type": "number",
            "title": "Progress Percentage"
          },
          "current_step": {
            "type": "string",
            "title": "Current Step"
          },
          "estimated_time_remaining": {
            "type": "number",
            "title": "Estimated Time Remaining"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "status",
          "progress_percentage",
          "current_step",
          "estimated_time_remaining",
          "details"
        ],
        "title": "ProcessingStatus",
        "description": "Processing status response"
      },
...
      "UploadResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "processing_results": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Results"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "progress": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "UploadResponse",
        "description": "Response model for upload endpoint"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}